+++ apps/storybook/vite.config.mjs
@@ -1,5 +1,5 @@
 import react from "@vitejs/plugin-react";
-import flowbiteReact from "flowbite-react/plugin/vite";
+import flowbiteReact from "@herob191/flowbite-react/plugin/vite";
 import { defineConfig } from "vite";
 
 export default defineConfig({